pyproject.toml
[tool.poetry]
name = "isabot"
version = "0.1.0"
description = ""
authors = ["John Carlo Manuel <39042556+johncmanuel@users.noreply.github.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.11"
fastapi = "^0.104.1"
uvicorn = {extras = ["standard"], version = "^0.24.0.post1"}
python-dotenv = "^1.0.0"
aiohttp = "^3.9.0"
pynacl = "^1.5.0"


[tool.poetry.group.dev.dependencies]
black = "^23.11.0"
flake8 = "^6.1.0"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
